backend/requirements.txt
fastapi==0.115.2
uvicorn==0.30.6
pydantic==2.9.2
email-validator==2.2.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
python-dotenv==1.0.1
pytest==8.3.3
httpx==0.27.2
pdfplumber==0.11.4
